+++ landingzones/templatetags/landingzones_tags.py
@@ -3,9 +3,6 @@
 from django import template
 from django.urls import reverse
 
-# Projectroles dependency
-from projectroles.plugins import get_backend_api
-
 from landingzones.constants import (
     STATUS_STYLES,
     STATUS_FINISHED,
@@ -69,21 +66,6 @@ def disable_zone_ui(zone, user):
     elif not user.is_superuser and zone.status in STATUS_ALLOW_UPDATE:
         return False
     return True
-
-
-@register.simple_tag
-def get_zone_list_url(zone):
-    """Return iRODS file list querying URL for landing zone"""
-    irods_backend = get_backend_api('omics_irods')
-    if not irods_backend:
-        return None
-    return irods_backend.get_url(
-        view='list',
-        project=zone.project,
-        path=irods_backend.get_path(zone),
-        md5=True,
-        colls=True,
-    )
 
 
 @register.simple_tag